+++ database/migrations/2024_09_21_021519_create_master_payment_term_table.php
@@ -12,7 +12,7 @@ return new class extends Migration
     public function up(): void
     {
         Schema::create('master_payment_term', function (Blueprint $table) {
-            $table->unsignedTinyInteger('id', 3);
+            $table->increments('id', 3);
             $table->string('term_name', 60);
             $table->double('term_amount')->default(0);
             $table->string('description')->nullable();
